LayoutTests/platform/glib/imported/w3c/web-platform-tests/html/capability-delegation/delegate-fullscreen-request-popup-cross-origin.https.sub.tentative-expected.txt

CONSOLE MESSAGE: TypeError: document.body.requestFullscreen is not a function. (In 'document.body.requestFullscreen()', 'document.body.requestFullscreen' is undefined)
Verifies that element.requestFullscreen() calls from a cross-origin popup without user activation work if and only if the opener has user activation and it delegates the capability. https://wicg.github.io/capability-delegation/spec.html

Harness Error (TIMEOUT), message = null

TIMEOUT Fullscreen requests from a cross-origin popup fails without delegation from an opener with no user activation Test timed out
NOTRUN Fullscreen requests from a cross-origin popup fails without delegation from an opener with user activation
NOTRUN Fullscreen requests from a cross-origin popup succeeds with delegation from an opener with user activation
